requirements.txt
aiohttp==3.8.1
python-dotenv==0.20.0
aiohttp-apispec==2.2.3
marshmallow==3.15.0
httpx==0.22.0
aiormq==6.2.3
requests
